In test change group of directory again
This has no effect on the test itself but the code may be used as an example where it does matter.
Since the systemd unit runs with DynamicUser=true, the group ID of the directory may change over time, so reset it.
This reverts part of 9a4e3bb9fa7a99bf4fc1d319806ed7388a3bee18.